igreja_sao_sebastiao/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "igreja-sao-sebastiao"
version = "2.0.0"
description = "Site da Igreja Sao Sebastiao - Ponte Nova/MG"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1,<6.0",
    "django-unfold>=0.40",
    "Pillow>=11.0",
    "gunicorn>=23.0",
    "whitenoise>=6.7",
    "python-dotenv>=1.0",
    "psycopg[binary]>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "django-debug-toolbar>=4.4",
]

[tool.setuptools.packages.find]
include = ["apps*", "config*"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["apps", "config"]
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "DJ"]
ignore = ["DJ001"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
]
